+++ include/datastructure/Route.hpp
@@ -56,7 +56,7 @@ public:
   [[nodiscard]] int get_edge(int route_index) const;
   [[nodiscard]] const cda_rail::Edge&
   get_edge(int route_index, const cda_rail::Network& network) const;
-  [[nodiscard]] int                     size() const { return edges.size(); };
+  [[nodiscard]] size_t                  size() const { return edges.size(); };
   [[nodiscard]] bool                    empty() const { return edges.empty(); };
   [[nodiscard]] const std::vector<int>& get_edges() const { return edges; };
 
@@ -121,7 +121,7 @@ public:
   [[nodiscard]] bool has_route(const std::string& train_name) const {
     return routes.find(train_name) != routes.end();
   };
-  [[nodiscard]] int          size() const { return routes.size(); };
+  [[nodiscard]] size_t       size() const { return routes.size(); };
   [[nodiscard]] bool         empty() const { return routes.empty(); };
   [[nodiscard]] const Route& get_route(const std::string& train_name) const;
